Modularize Environment logic
