+++ crates/vm/levm/src/lib.rs
@@ -2,6 +2,7 @@ pub mod account;
 pub mod call_frame;
 pub mod constants;
 pub mod db;
+pub mod environment;
 pub mod errors;
 pub mod memory;
 pub mod opcode_handlers;
@@ -11,3 +12,4 @@ pub mod utils;
 pub mod vm;
 
 pub use account::*;
+pub use environment::*;
