+++ crates/vm/levm/src/vm.rs
@@ -3,6 +3,7 @@ use crate::{
     call_frame::CallFrame,
     constants::*,
     db::{Cache, Database},
+    environment::Environment,
     errors::{OpcodeSuccess, ResultReason, TransactionReport, TxResult, VMError},
     opcodes::Opcode,
 };
@@ -21,47 +22,6 @@ pub struct Substate {
     // accessed addresses and storage keys are considered WARM
     // pub accessed_addresses: HashSet<Address>,
     // pub accessed_storage_keys: HashSet<(Address, U256)>,
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct Environment {
-    /// The sender address of the transaction that originated
-    /// this execution.
-    pub origin: Address,
-    pub consumed_gas: U256,
-    pub refunded_gas: U256,
-    pub gas_limit: U256,
-    pub block_number: U256,
-    pub coinbase: Address,
-    pub timestamp: U256,
-    pub prev_randao: Option<H256>,
-    pub chain_id: U256,
-    pub base_fee_per_gas: U256,
-    pub gas_price: U256,
-    pub block_excess_blob_gas: Option<U256>,
-    pub block_blob_gas_used: Option<U256>,
-    pub tx_blob_hashes: Option<Vec<H256>>,
-}
-
-impl Environment {
-    pub fn default_from_address(origin: Address) -> Self {
-        Self {
-            origin,
-            consumed_gas: TX_BASE_COST,
-            refunded_gas: U256::zero(),
-            gas_limit: U256::MAX,
-            block_number: Default::default(),
-            coinbase: Default::default(),
-            timestamp: Default::default(),
-            prev_randao: Default::default(),
-            chain_id: U256::one(),
-            base_fee_per_gas: Default::default(),
-            gas_price: Default::default(),
-            block_excess_blob_gas: Default::default(),
-            block_blob_gas_used: Default::default(),
-            tx_blob_hashes: Default::default(),
-        }
-    }
 }
 
 pub struct VM {
